refactor: reorganize component imports and remove deprecated proxy references
