+++ packages/studiocms/src/componentRegistry/convert-hyphens.ts
