+++ packages/studiocms/src/components/renderers/studiocms-html.astro
@@ -1,5 +1,5 @@
 ---
-import { importComponentsKeys } from '../../lib/renderer/runtime.js';
+import { importComponentsKeys } from '../../componentRegistry/runtime.js';
 import { shared } from '../../lib/renderer/shared.js';
 import { createComponentProxy, transformHTML } from '../../runtime/AstroComponentProxy.js';
 import type { PluginPageTypeRendererProps } from '../../types.js';
