+++ packages/studiocms/src/components/renderers/studiocms-markdown.astro
@@ -1,6 +1,6 @@
 ---
 import { mdPreRender } from 'studiocms:plugins/renderers';
-import { importComponentsKeys } from '../../lib/renderer/runtime.js';
+import { importComponentsKeys } from '../../componentRegistry/runtime.js';
 import { shared } from '../../lib/renderer/shared.js';
 import { createComponentProxy, transformHTML } from '../../runtime/AstroComponentProxy.js';
 import type { PluginPageTypeRendererProps } from '../../types.js';
